requirements.txt
certifi==2018.1.18
chardet==3.0.4
Django==1.11.6
django-forms-bootstrap==3.1.0
idna==2.6
pytz==2017.2
requests==2.18.4
stripe==1.77.1
urllib3==1.22
mysql-python==1.2.5
gunicorn==19.7.1
dj-database-url==0.2.1
whitenoise==3.3.1
django-recaptcha==1.4.0